Introduce ReplicationCoordinator to support multiple AZs
It's common for applications that are deployed across multiple
availability zones (using a replicated database) to create an ad-hoc
method for processes to discover which zone is "active", meaning the
zone primarily responsible for writing to the database.

For example, a team may choose to use a MySQL system variable to
indicate the data center where the primary database sits. In which
case, they need to write code to make sure all Rails processes in all
zones query this efficiently (it may be slow to access in non-primary
zones) and are notified if the primary zone changes, as in the case of
a data center failover.

`ReplicationCoordinator::Base` is introduced to allow developers to
write code that determines whether a process is in an active zone, and
then:

- monitor and cache that value, with configurable polling interval
- fire callbacks when the state changes from active -> passive or vice versa

Additionally, a test helper class is provided to simplify testing
failover behavior.

Finally, Rails is configured by default to use a simple concrete
replication coordinator, `SingleZone`, which always indicates the
caller is in an active zone.